rename PartnerReferral type to SubmittedLead and update related templates
