+++ apps/web/lib/zod/schemas/rewards.ts
@@ -112,7 +112,7 @@ export const REWARD_CONDITIONS: Record<
               },
               {
                 id: "submitted",
-                label: "submitted referral",
+                label: "submitted lead",
               },
               {
                 id: "trial",
